+++ includes/output/Image.php
@@ -17,7 +17,18 @@ class Image extends Output
     'description' => 'Output in image format. The data fed into the output can be a URL (must start with http) or an input filename.',
     'menu' => 'Output',
     'application' => 'All',
-    'input' => array(),
+    'input' => array(
+      'destination' => array(
+        'description' => 'A single or array of URLs to send the results to.',
+        'cardinality' => array(1, '*'),
+        'accepts' => array('processor', 'literal'),
+      ),
+      'options' => array(
+        'description' => 'Extra Curl options to be applied when sent to the destination  (e.g. cursor: -1, screen_name: foobarapi, skip_status: true, etc).',
+        'cardinality' => array(0, '*'),
+        'accepts' => array('processor field'),
+      ),
+    ),
   );
 
   protected function getData()