+++ includes/output/Json.php
@@ -11,7 +11,18 @@ class Json extends Output
     'description' => 'Output in JSON format.',
     'menu' => 'Output',
     'application' => 'All',
-    'input' => array(),
+    'input' => array(
+      'destination' => array(
+        'description' => 'A single or array of URLs to send the results to.',
+        'cardinality' => array(1, '*'),
+        'accepts' => array('processor', 'literal'),
+      ),
+      'options' => array(
+        'description' => 'Extra Curl options to be applied when sent to the destination  (e.g. cursor: -1, screen_name: foobarapi, skip_status: true, etc).',
+        'cardinality' => array(0, '*'),
+        'accepts' => array('processor field'),
+      ),
+    ),
   );
 
   /**